Added an individual posting page
Added a new component to pages which is the individual posting page. There is now a draft of how it should look like in the html file and most of the css is inline and very few code in the css file. In the future it will have to be cleaned up. the .spec.ts and .ts files are left untouched besides what is autogenerated. In addition, i had an error when added this component to pages. therefore, i had to import the RouterModule in app.module.ts   in two places.